requirement.txt
Django==3.2.5
django-simple-captcha==0.5.14
django-ckeditor==6.1.0
django-anymail==8.4
django-widget-tweaks==1.4.11